+++ index.js
@@ -5,11 +5,12 @@
  */
 "use strict";
 
-var client = require('./src/client'),
-    data = require('./src/data'),
-    replace = require('./src/graphql/replace'),
-    exprParser = require('./src/graphql/exprParser'),
-    find = require('./src/graphql/find');
+const client = require('./src/client');
+const data = require('./src/data');
+const replace = require('./src/graphql/replace');
+const exprParser = require('./src/graphql/exprParser');
+const find = require('./src/graphql/find');
+const HaystackTools = require('./src/utils/haystack');
 
 /* Exported symbols */
 const jsWidesky = {
@@ -34,7 +35,9 @@ const jsWidesky = {
         replace,
         exprParser,
         find
-    }
+    },
+    // Haystack tools
+    hsTools: HaystackTools
 };
 
 module.exports = jsWidesky;